+++ examples/Example/HayaEcomm.cs
@@ -1,5 +1,10 @@
 ﻿using Haya;
-[assembly:Meta(AppName = "CheckoutApi", Description = "Handles processing payment for a shopping cart", System = "HayaEcomm")]
+[assembly:Meta(
+    AppName = "PaymentsApi", 
+    Description = "Handles processing payment for a shopping cart", 
+    System = "Ordering",
+    Team = "Ordering Team",
+    Repository = "org/payment-api")]
 
 namespace Example;
 
@@ -12,16 +17,21 @@ public class CreditCardPaymentUsecase
     Direction = Direction.Upstream,
     System = "WebShop",
     Description = "Handles cart checkout for a customer",
-    DataDescription = "Request payment")]
+    Tech = "TypeScript",
+    Protocol = "HTTPS",
+    DataDescription = "Request payment",
+    Repository = "org/checkout-api")]
 public class PaymentsController
 { }
 
 [Collaborator(AppName = "Visa Payment Provider",
+    Protocol = "HTTPS",
     Relationship = Relationship.External,
     Direction = Direction.Downstream,
     System = "VISA",
     Description = "Handles payment processing of Visa cards",
-    DataDescription = "Process payment")]
+    DataDescription = "Process payment",
+    Repository = "org/webshop")]
 public class VisaPaymentService
 { }
 
@@ -30,6 +40,9 @@ public class VisaPaymentService
     Direction = Direction.Downstream,
     System = "Shipping",
     Description = "Handles shipping of orders",
-    DataDescription = "Ship order")]
+    Tech = "dotnet",
+    Protocol = "MQTT",
+    DataDescription = "Ship order",
+    Repository = "org/shipping-service")]
 public class ShippingService
 { }